split test files in integration and unit tests
